Add queryParam support for show entities
Adds a mixin that will translate the filterData properties into a comma
separated queryParam (show=). At the moment it is used by the
project-version controller since it is the parent of all the controllers
that would need the filterData. It is started as a mixin because I
thought it had to be part of each controller but (Ember being
cool) it can also work in one spot in the parent hierarchy. I left it as
a mixin simple because the project-version controller is almost 100
lines large so far.

I aded one test to see the query params show up when you click the check
boxes. I also had to fix some tests that had to brittle expectations of
the URL. I simply stripped out the query params from the URL since the
tests were not asserting anything regarding params.