search-input: update to glimmer
Let's update the SearchInput component to extend glimmer/component
instead of ember/component. Glimmer components offer a bunch of various
advantages over classic components and are the direction that ember is
heading in general.

While we're here, let's also re-enable the eslint rules for which the
SearchInput component had the only violations.