Replace scroll-tracker mixin with class decorator
NOTE: there may be an issue with this on FastBoot (there is a report to
that effect in the output in the console). It is not clear *why*,
though, since the behavior in this decorated class *should* be the same
as in the mixin. Possibly the difference between the route events and
the router events is coming into play?